requirements.txt
streamlit
langchain
langchain-groq
selenium 
beautifulsoup4
lxml
html5lib
python-dotenv